Extract run_execution to its own module to fix God function smell.
